requirements.txt
Flask>=3.0.0
Flask-Limiter>=3.5.0
oracledb>=2.0.0
requests>=2.28.0
flask-socketio>=5.3.0
python-socketio>=5.10.0
eventlet>=0.33.0
psutil>=6.0.0
selenium>=4.0.0
Flask-Babel>=4.0.0
Babel>=2.13.0
python-dotenv>=1.0.0
markdown>=3.10.0
openpyxl>=3.1.0
reportlab>=4.0.0
mammoth>=1.6.0
pdfminer.six>=20221105

